% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcgp.R
\name{bcgp}
\alias{bcgp}
\title{Performing Bayesian (Composite) Gaussian Processes}
\usage{
\code{bcgp(trainingX, trainingY, ...)}
}
\arguments{
\item{trainingX}{An \emph{n x d} matrix of input locations}

\item{trainingY}{An \emph{n x 1} vector of responses}
}
\value{
An S4 method I hope. Who the fuck knows.
}
\description{
\code{bcgp} is the main function in this package. The user
will be able to choose any combination of prediction/variable
selection, stationary/non-stationary, and composite/standard 
Gaussian Processes
}
\details{
I'll edit this as the package comes together
}
\examples{
trainingX <- matrix(rnorm(5 * 2), ncol = 2, nrow = 5)
trainingY <- rnorm(5, 10, 3)
bcgp(trainingX, trainingY)
}
\seealso{
\code{\link{rescale}}
}
\author{
Casey Davis (\email{cbdavis33@gmail.com})
}
