\name{bcgpRcpp-package}
\alias{bcgpRcpp-package}
\alias{bcgpRcpp}
\docType{package}
\title{
  Bayesian Composite Gaussian Processes
}
\description{
  This package implements the work in "A Bayesian Approach to 
  Prediction and Variable Selection Using Non-Stationary Composite 
  Gaussian Processes".
}
\details{
  This section should provide a more detailed overview of how to use the
  package, including the most important functions.
}
\author{
Casey Davis

Maintainer: Casey Davis <cbdavis33@gmail.com>
}
\references{
  Reference a couple papers or my dissertation
}
\keyword{ package }
\seealso{
  Optional links to other man pages
}
\examples{
  \dontrun{
     ## Optional simple examples of the most important functions
     ## These can be in \dontrun{} and \donttest{} blocks.   
  }
}
