% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stdX.R
\name{stdX}
\alias{stdX}
\title{Rescale each column of a matrix to [0, 1]}
\usage{
stdX(x)
}
\arguments{
\item{\code{x}}{A matrix input}
}
\value{
A matrix with columns rescaled to [0, 1]
}
\description{
Rescale each column of a matrix to [0, 1]
}
\details{
Rescale each column of a matrix to [0, 1]
}
\examples{
stdX(matrix(rnorm(5 * 2), ncol = 2, nrow = 5))
}
\seealso{
\code{\link{rescale}}
}
\author{
Casey Davis (\email{cbdavis33@gmail.com})
}
